Performance Update
Experimented with @btime and @profiler to increase performance. Removing the scalar flux derivative tallies significantly increases performance. In total the code is 4x faster per sweep and uses 1/3 of the memory.

Of course there may be times when we want to tally dphi so may need to come up with a way to turn the tally on later.

Still need to experiment with @views for additional performance enhancements.